split chatapp into 2 components since we slowly have too much logic inside it
